% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter low-expressed genes}
\usage{
filter_genes(data, threshold = 10)
}
\arguments{
\item{data}{A dataframe containing gene expression data.}

\item{threshold}{Numeric value for minimum expression threshold. Default is 10.}
}
\value{
A dataframe with filtered genes.
}
\description{
This function filters genes with low expression from the dataset.
}
